move commander functions to UserProfileCommander
