+++ kifi-backend/modules/shoebox/test/com/keepit/commanders/LibraryCommanderTest.scala
@@ -50,18 +50,6 @@ class LibraryCommanderTest extends TestKitSupport with SpecificationLike with Sh
     FakeHeimdalServiceClientModule()
   )
 
-  private val profileLibraryOrdering = new Ordering[Library] {
-    def compare(self: Library, that: Library): Int =
-      (self.kind compare that.kind) match {
-        case 0 =>
-          (self.memberCount compare that.memberCount) match {
-            case 0 => -(self.id.get.id compare that.id.get.id)
-            case c => -c
-          }
-        case c => c
-      }
-  }
-
   def setupUsers()(implicit injector: Injector) = {
     val t1 = new DateTime(2014, 7, 4, 12, 0, 0, 0, DEFAULT_DATE_TIME_ZONE)
     val emailRepo = inject[UserEmailAddressRepo]
@@ -1196,120 +1184,6 @@ class LibraryCommanderTest extends TestKitSupport with SpecificationLike with Sh
       }
     }
 
-    "get invited to libraries" in {
-      withDb(modules: _*) { implicit injector =>
-        implicit val config = inject[PublicIdConfiguration]
-        val libraryCommander = inject[LibraryCommander]
-        val (user1, user2, user3, libs) = db.readWrite { implicit s =>
-          val user1 = user().saved
-          val user2 = user().saved
-          val user3 = user().saved
-          val lib1 = libraries(2).map(_.published().withUser(user1)).saved.head.savedInvitation(user2)
-          val lib2 = libraries(2).map(_.secret().withUser(user1)).saved.head.savedInvitation(user2).savedInvitation(user3)
-          val lib3 = libraries(2).map(_.published().withUser(user2)).saved.head.savedInvitation(user1)
-          invite().fromLibraryOwner(library.published().withUser(user2).saved).saved.savedStateChange(LibraryInviteStates.ACCEPTED)
-          invite().fromLibraryOwner(library.published().withUser(user2).saved).declined.saved
-          libraries(10).map(_.published().withUser(user1)).saved
-          libraries(10).map(_.published().withUser(user2)).saved
-          (user1, user2, user3, lib1 :: lib2 :: lib3 :: Nil)
-        }
-
-        libraryCommander.getInvitedLibraries(user1, None, Paginator(0, 5), ImageSize("100x100")).size === 0
-        libraryCommander.getInvitedLibraries(user1, Some(user2), Paginator(0, 5), ImageSize("100x100")).size === 0
-
-        val libs1 = libraryCommander.getInvitedLibraries(user1, Some(user1), Paginator(0, 5), ImageSize("100x100"))
-        libs1.size === 1
-        libs1.map(_.id).head === Library.publicId(libs(2).id.get)
-
-        val libs2 = libraryCommander.getInvitedLibraries(user2, Some(user2), Paginator(0, 5), ImageSize("100x100"))
-        libs2.size === 2
-        libs2.map(_.id) === libs.take(2).reverse.map(_.id.get).map(Library.publicId)
-
-        val libs3 = libraryCommander.getInvitedLibraries(user3, Some(user3), Paginator(0, 5), ImageSize("100x100"))
-        libs3.size === 1
-        libs3.map(_.id).head === Library.publicId(libs(1).id.get)
-
-      }
-    }
-
-    "getFollowingLibraries for anonymous and paginate" in {
-      withDb(modules: _*) { implicit injector =>
-        implicit val config = inject[PublicIdConfiguration]
-        val libraryCommander = inject[LibraryCommander]
-        val (owner, other, allLibs) = db.readWrite { implicit s =>
-          val owner = user().saved
-          val other = user().saved
-          val otherFollows1 = libraries(2).map(_.published().withUser(owner)).saved.head.savedFollowerMembership(other)
-          val otherFollows2 = libraries(2).map(_.secret().withUser(owner)).saved.head.savedFollowerMembership(other)
-          libraries(2).map(_.published().withUser(other)).saved.head.savedFollowerMembership(owner)
-          val otherFollows3 = libraries(10).map(_.published().withUser(owner)).saved.map(_.savedFollowerMembership(other))
-          libraries(10).map(_.published().withUser(other)).saved
-          (owner, other, List(otherFollows1) ++ otherFollows3)
-        }
-
-        val libsP1 = libraryCommander.getFollowingLibraries(other, None, Paginator(0, 5), ImageSize("100x100"))
-        libsP1.size === 5
-        libsP1.map(_.id) === allLibs.reverse.take(5).map(_.id.get).map(Library.publicId)
-
-        val libsP2 = libraryCommander.getFollowingLibraries(other, None, Paginator(1, 5), ImageSize("100x100"))
-        libsP2.size === 5
-        libsP2.map(_.id) === allLibs.reverse.drop(5).take(5).map(_.id.get).map(Library.publicId)
-
-        val libsP3 = libraryCommander.getFollowingLibraries(other, None, Paginator(2, 5), ImageSize("100x100"))
-        libsP3.size === 1
-        libsP3.map(_.id) === allLibs.reverse.drop(10).take(5).map(_.id.get).map(Library.publicId)
-      }
-    }
-
-    "getFollowingLibraries for self" in {
-      withDb(modules: _*) { implicit injector =>
-        implicit val config = inject[PublicIdConfiguration]
-        val libraryCommander = inject[LibraryCommander]
-        val (owner, other, allLibs) = db.readWrite { implicit s =>
-          val owner = user().saved
-          val other = user().saved
-          val otherFollows1 = libraries(2).map(_.published().withUser(owner)).saved.head.savedFollowerMembership(other)
-          val otherFollows2 = libraries(2).map(_.secret().withUser(owner)).saved.head.savedFollowerMembership(other)
-          libraries(2).map(_.published().withUser(other)).saved.head.savedFollowerMembership(owner)
-          val otherFollows3 = libraries(10).map(_.published().withUser(owner)).saved.map(_.savedFollowerMembership(other))
-          libraries(10).map(_.published().withUser(other)).saved
-          (owner, other, List(otherFollows1, otherFollows2) ++ otherFollows3)
-        }
-
-        val libs = libraryCommander.getFollowingLibraries(other, Some(other), Paginator(0, 500), ImageSize("100x100"))
-        libs.size === 12
-        libs.head.owner.externalId === owner.externalId
-        libs.map(_.id) === allLibs.reverse.map(_.id.get).map(Library.publicId)
-      }
-    }
-
-    "getFollowingLibraries for other" in {
-      withDb(modules: _*) { implicit injector =>
-        implicit val config = inject[PublicIdConfiguration]
-        val libraryCommander = inject[LibraryCommander]
-        val (user1, user2, user3, follows1, follows2, follows3, follows4) = db.readWrite { implicit s =>
-          val user1 = user().saved
-          val user2 = user().saved
-          val user3 = user().saved
-          val follows1 = libraries(2).map(_.published().withUser(user1).withName("a")).saved.head.savedFollowerMembership(user2)
-          val follows2 = libraries(2).map(_.secret().withUser(user1).withName("b")).saved.head.savedFollowerMembership(user2).savedFollowerMembership(user3)
-          val follows3 = libraries(10).map(_.secret().withUser(user3).withName("c")).saved.map(_.savedFollowerMembership(user2))
-          libraries(2).map(_.published().withUser(user2).withName("d")).saved.head.savedFollowerMembership(user1)
-          val follows4 = libraries(10).map(_.published().withUser(user1).withName("e")).saved.map(_.savedFollowerMembership(user2))
-          libraries(10).map(_.published().withUser(user2).withName("f")).saved
-          (user1, user2, user3, follows1, follows2, follows3, follows4)
-        }
-
-        val self = libraryCommander.getFollowingLibraries(user2, Some(user2), Paginator(0, 500), ImageSize("100x100"))
-        self.size === 22
-        self.map(_.id) === (List(follows1, follows2) ++ follows3 ++ follows4).reverse.map(_.id.get).map(Library.publicId)
-
-        val libs1 = libraryCommander.getFollowingLibraries(user2, Some(user1), Paginator(0, 500), ImageSize("100x100"))
-        libs1.size === 12
-        libs1.map(_.id) === (List(follows1, follows2) ++ follows4).reverse.map(_.id.get).map(Library.publicId)
-      }
-    }
-
     "sortFollowers" in {
       withDb(modules: _*) { implicit injector =>
         val libraryCommander = inject[LibraryCommander]
@@ -1320,130 +1194,6 @@ class LibraryCommanderTest extends TestKitSupport with SpecificationLike with Sh
         val user5 = BasicUser.fromUser(user().withPictureName("b").get)
         val sorted = libraryCommander.sortUsersByImage(user1 :: user2 :: user3 :: user4 :: user5 :: Nil)
         sorted.map(_.pictureName) === (user2 :: user4 :: user5 :: user1 :: user3 :: Nil).map(_.pictureName)
-      }
-    }
-
-    "get ownerLibraries for friend" in {
-      withDb(modules: _*) { implicit injector =>
-        implicit val config = inject[PublicIdConfiguration]
-        val libraryCommander = inject[LibraryCommander]
-        val (owner, other, friend, allLibs) = db.readWrite { implicit s =>
-          val owner = user().saved
-          val other = user().saved
-          val friend = user().saved
-          val ownerLibs1 = libraries(2).map(_.published().withUser(owner)).saved
-          library.secret().withUser(owner).saved
-          val ownerPrivLib = library.secret().withUser(owner).saved.savedFollowerMembership(friend)
-          libraries(2).map(_.published().withUser(other)).saved
-          libraries(2).map(_.secret().withUser(other)).saved
-          val ownerLibs2 = libraries(10).map(_.published().withUser(owner)).saved
-          libraries(10).map(_.published().withUser(other)).saved
-
-          libraryRepo.all.map { lib =>
-            keeps(2).map(_.withLibrary(lib)).saved
-          }
-          (owner, other, friend, ownerLibs1 ++ List(ownerPrivLib) ++ ownerLibs2)
-        }
-
-        libraryCommander.getOwnProfileLibraries(owner, None, Paginator(0, 1000), ImageSize("100x100")).size === 12
-
-        val libsForOther = libraryCommander.getOwnProfileLibraries(owner, Some(other), Paginator(0, 1000), ImageSize("100x100"))
-        libsForOther.size === 12
-
-        val libsForFriend = libraryCommander.getOwnProfileLibraries(owner, Some(friend), Paginator(0, 1000), ImageSize("100x100"))
-        libsForFriend.size === 13
-        libsForFriend.map(_.id) === allLibs.reverse.map(_.id.get).map(Library.publicId)
-      }
-    }
-
-    "get ownerLibraries for self" in {
-      withDb(modules: _*) { implicit injector =>
-        implicit val config = inject[PublicIdConfiguration]
-        val libraryCommander = inject[LibraryCommander]
-        val (owner, other, allLibs) = db.readWrite { implicit s =>
-          val owner = user().saved
-          val other = user().saved
-          val ownerLibs1 = libraries(2).map(_.published().withUser(owner)).saved
-          val ownerPrivLib = library.secret().withUser(owner).saved
-          libraries(2).map(_.published().withUser(other)).saved
-          libraries(2).map(_.secret().withUser(other)).saved
-          val ownerLibs2 = libraries(10).map(_.published().withUser(owner)).saved
-          libraries(10).map(_.published().withUser(other)).saved
-
-          libraryRepo.all.take(4).map { lib =>
-            keeps(2).map(_.withLibrary(lib)).saved
-          }
-
-          (owner, other, ownerLibs1 ++ List(ownerPrivLib) ++ ownerLibs2)
-        }
-
-        // 4 libs with keeps. First two are private, next two should be seen by anonymous.
-        libraryCommander.getOwnProfileLibraries(owner, None, Paginator(0, 1000), ImageSize("100x100")).size === 2
-
-        val libsForOther = libraryCommander.getOwnProfileLibraries(owner, Some(other), Paginator(0, 1000), ImageSize("100x100"))
-        libsForOther.size === 2
-
-        val libsForFriend = libraryCommander.getOwnProfileLibraries(owner, Some(owner), Paginator(0, 1000), ImageSize("100x100"))
-        libsForFriend.size === 3
-      }
-    }
-
-    "get ownerLibraries for anonymous and paginate" in {
-      withDb(modules: _*) { implicit injector =>
-        implicit val config = inject[PublicIdConfiguration]
-        val libraryCommander = inject[LibraryCommander]
-        val (owner, allLibs) = db.readWrite { implicit s =>
-          val owner = user().saved
-          val other = user().saved
-          val ownerLibs1 = libraries(2).map(_.published().withUser(owner)).saved
-          libraries(2).map(_.secret().withUser(owner)).saved
-          libraries(2).map(_.published().withUser(other)).saved
-          val ownerLibs2 = libraries(10).map(_.published().withUser(owner).withMemberCount(6)).saved
-          libraries(10).map(_.published().withUser(other).withMemberCount(6)).saved
-          val ownerLibs3 = libraries(3).map(_.published().withUser(owner)).saved
-          libraries(3).map(_.published().withUser(other)).saved
-
-          libraryRepo.all.map { lib =>
-            keeps(2).map(_.withLibrary(lib)).saved
-          }
-
-          (owner, ownerLibs1 ++ ownerLibs2 ++ ownerLibs3)
-        }
-
-        val ord = profileLibraryOrdering
-
-        val libsP1 = libraryCommander.getOwnProfileLibraries(owner, None, Paginator(0, 5), ImageSize("100x100"))
-        libsP1.size === 5
-        libsP1.map(_.id) === allLibs.sorted(ord).take(5).map(_.id.get).map(Library.publicId)
-
-        val libsP2 = libraryCommander.getOwnProfileLibraries(owner, None, Paginator(1, 5), ImageSize("100x100"))
-        libsP2.size === 5
-        libsP2.map(_.id) === allLibs.sorted(ord).drop(5).take(5).map(_.id.get).map(Library.publicId)
-
-        val libsP3 = libraryCommander.getOwnProfileLibraries(owner, None, Paginator(2, 5), ImageSize("100x100"))
-        libsP3.size === 5
-        libsP3.map(_.id) === allLibs.sorted(ord).drop(10).take(5).map(_.id.get).map(Library.publicId)
-      }
-    }
-
-    "get owner library counts" in {
-      withDb(modules: _*) { implicit injector =>
-        val libraryCommander = inject[LibraryCommander]
-        db.readWrite { implicit s =>
-          val libs = libraries(5)
-          val user1 = libs.take(3).map { _.withUser(Id[User](1)).published() }.saved
-          val user2 = libs.drop(3).map { _.withUser(Id[User](2)).published() }.saved
-        }
-
-        db.readOnlyMaster { implicit s =>
-          val users = Set(1, 2, 100).map { Id[User](_) }
-          val map = libraryRepo.getOwnerLibraryCounts(users)
-          map === Map(Id[User](1) -> 3, Id[User](2) -> 2, Id[User](100) -> 0)
-
-          val map2 = libraryRepo.getOwnerLibraryCounts(Set[Id[User]]())
-          map2.size === 0
-        }
-
       }
     }
